requirements.txt
aiofiles==24.1.0
aiohttp==3.9.3
aiopath==0.6.11
beautifulsoup4==4.13.4
browser_use==0.2.5
celery==5.4.0
chardet==4.0.0
dashscope==1.23.1
docx==0.2.4
duckduckgo_search==7.5.1
edge_tts==7.0.0
fastapi==0.115.12
fitz==0.0.1.dev2
langchain==0.3.25
langchain_openai==0.3.18
numpy==2.2.6
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pdfplumber==0.11.5
Pillow==11.2.1
playwright==1.51.0
pydantic==2.11.5
pydub==0.25.1
pymilvus==2.5.5
python_docx==1.1.2
redis==5.2.1
Requests==2.32.3
sentence_transformers==3.4.1
simpleaudio==1.0.4
spacy==3.8.4
speechrecognition==3.14.2
SQLAlchemy==2.0.25
starlette==0.46.2
torch==2.2.2
